+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### ⬆️ Improved
 
 ### ✅ Added
-
+- Adding ChatDomain.connectionState from with 3 states: CONNECTED, CONNECTING and OFFLINE. Also changing the exibition of disconnected state in ChannelListHeaderView and MessageListHeaderView. Please use this instead of ChatDomain.online
 ### ⚠️ Changed
 
 ### ❌ Removed
@@ -63,6 +63,7 @@
 ### ⬆️ Improved
 
 ### ✅ Added
+- Added `streamUiMessageMaxPossibleWidthFactorMine` and `streamUiMessageMaxPossibleWidthFactorTheirs` `MessageListView` attributes. You can make messages wider by passing values < 100%.
 
 ### ⚠️ Changed